fix code check

🤖 AI[0%] 🔧 Human Fixed[0%] 🧑 Human[100%] 👌 AI Adopted[0%]

+++ services/distributeddataservice/service/data_share/subscriber_managers/published_data_subscriber_manager.cpp
@@ -46,7 +46,7 @@ int PublishedDataSubscriberManager::Add(const PublishedDataKey &key,
                     node.enabled = true;
                     node.isNotifyOnEnabled = false;
                     node.callerTokenId = IPCSkeleton::GetCallingTokenID();
-                    node.callerPid = IPCSkeleton::GetCallingPid();
+                    node.callerPid = static_cast<uint32_t>(IPCSkeleton::GetCallingPid());
                     node.userId = userId;
                     return true;
                 }
